requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain-openai==0.0.5
openai>=1.14.0
pydantic==2.4.2
pydantic-settings>=2.0.0
typing-extensions>=4.5.0
python-multipart>=0.0.6
langchain==0.0.352